test: split marketplace service regression tests
